Allow to modify symboltables in a constfst
