+++ rustfst/src/fst_impls/const_fst/fst.rs
@@ -14,6 +14,22 @@ impl<W: Semiring + 'static> Fst for ConstFst<W> {
     fn output_symbols(&self) -> Option<Rc<SymbolTable>> {
         self.osymt.clone()
     }
+
+    fn set_input_symbols(&mut self, symt: Rc<SymbolTable>) {
+        self.isymt = Some(Rc::clone(&symt))
+    }
+
+    fn set_output_symbols(&mut self, symt: Rc<SymbolTable>) {
+        self.osymt = Some(Rc::clone(&symt));
+    }
+
+    fn unset_input_symbols(&mut self) -> Option<Rc<SymbolTable>> {
+        self.isymt.take()
+    }
+
+    fn unset_output_symbols(&mut self) -> Option<Rc<SymbolTable>> {
+        self.osymt.take()
+    }
 }
 
 impl<W: Semiring> CoreFst for ConstFst<W> {
