+++ rustfst/src/fst_impls/vector_fst/fst.rs
@@ -16,6 +16,22 @@ impl<W: 'static + Semiring> Fst for VectorFst<W> {
     fn output_symbols(&self) -> Option<Rc<SymbolTable>> {
         self.osymt.clone()
     }
+
+    fn set_input_symbols(&mut self, symt: Rc<SymbolTable>) {
+        self.isymt = Some(Rc::clone(&symt))
+    }
+
+    fn set_output_symbols(&mut self, symt: Rc<SymbolTable>) {
+        self.osymt = Some(Rc::clone(&symt));
+    }
+
+    fn unset_input_symbols(&mut self) -> Option<Rc<SymbolTable>> {
+        self.isymt.take()
+    }
+
+    fn unset_output_symbols(&mut self) -> Option<Rc<SymbolTable>> {
+        self.osymt.take()
+    }
 }
 
 impl<W: 'static + Semiring> CoreFst for VectorFst<W> {
