+++ rustfst/src/fst_impls/vector_fst/mutable_fst.rs
@@ -7,8 +7,7 @@ use crate::fst_impls::vector_fst::{VectorFst, VectorFstState};
 use crate::fst_traits::MutableFst;
 use crate::fst_traits::{CoreFst, MutableArcIterator};
 use crate::semirings::Semiring;
-use crate::{Arc, StateId, SymbolTable};
-use std::rc::Rc;
+use crate::{Arc, StateId};
 
 #[inline]
 fn equal_arc<W: Semiring>(arc_1: &Arc<W>, arc_2: &Arc<W>) -> bool {
@@ -257,22 +256,6 @@ impl<W: 'static + Semiring> MutableFst for VectorFst<W> {
         }
         arcs.truncate(n_arcs);
         // Truncate doesn't modify the capacity of the vector. Maybe a shrink_to_fit ?
-    }
-
-    fn set_input_symbols(&mut self, symt: Rc<SymbolTable>) {
-        self.isymt = Some(Rc::clone(&symt))
-    }
-
-    fn set_output_symbols(&mut self, symt: Rc<SymbolTable>) {
-        self.osymt = Some(Rc::clone(&symt));
-    }
-
-    fn unset_input_symbols(&mut self) -> Option<Rc<SymbolTable>> {
-        self.isymt.take()
-    }
-
-    fn unset_output_symbols(&mut self) -> Option<Rc<SymbolTable>> {
-        self.osymt.take()
     }
 }
