+++ rustfst/src/fst_traits/fst.rs
@@ -191,4 +191,31 @@ pub trait Fst:
     /// Retrieves the output `SymbolTable` associated to the Fst.
     /// If no SymbolTable has been previously attached then `None` is returned.
     fn output_symbols(&self) -> Option<Rc<SymbolTable>>;
+
+    /// Attaches an output `SymbolTable` to the Fst.
+    /// The `SymbolTable` is not duplicated with the use of Rc.
+    fn set_input_symbols(&mut self, symt: Rc<SymbolTable>);
+
+    /// Attaches an output `SymbolTable` to the Fst.
+    /// The `SymbolTable` is not duplicated with the use of Rc.
+    fn set_output_symbols(&mut self, symt: Rc<SymbolTable>);
+
+    /// Removes the input symbol table from the Fst and retrieves it.
+    fn unset_input_symbols(&mut self) -> Option<Rc<SymbolTable>>;
+    /// Removes the output symbol table from the Fst and retrieves it.
+    fn unset_output_symbols(&mut self) -> Option<Rc<SymbolTable>>;
+
+    fn set_symts_from_fst<OF: Fst>(&mut self, other_fst: &OF) {
+        if let Some(symt) = other_fst.input_symbols() {
+            self.set_input_symbols(symt);
+        } else {
+            self.unset_input_symbols();
+        }
+
+        if let Some(symt) = other_fst.output_symbols() {
+            self.set_output_symbols(symt);
+        } else {
+            self.unset_output_symbols();
+        }
+    }
 }
