+++ rustfst/src/tests_openfst/fst_impls/const_fst.rs
@@ -4,7 +4,7 @@ use failure::Fallible;
 
 use crate::algorithms::fst_convert_from_ref;
 use crate::fst_impls::{ConstFst, VectorFst};
-use crate::fst_traits::{Fst, MutableFst};
+use crate::fst_traits::Fst;
 use crate::semirings::{SerializableSemiring, WeightQuantize};
 use crate::tests_openfst::FstTestData;
 use crate::SymbolTable;
